requirements.txt
Flask==3.0.0
tensorflow==2.15.0
numpy
Pillow
opencv-python-headless
gunicorn
